rename split_triangle.c to .h
