+++ applications/MeshingApplication/custom_utilities/local_refine_triangle_mesh.hpp
@@ -30,7 +30,7 @@
 #include "geometries/line_2d_2.h"
 #include "geometries/line_3d_2.h"
 #include "custom_utilities/local_refine_geometry_mesh.hpp"
-#include "utilities/split_triangle.c"
+#include "utilities/split_triangle.h"
 
 namespace Kratos
 {
