+++ applications/TrilinosApplication/custom_utilities/trilinos_refine_mesh.h
@@ -17,7 +17,7 @@
 /* Project includes */
 #include "includes/deprecated_variables.h"
 #include "utilities/math_utils.h"
-#include "utilities/split_triangle.c"
+#include "utilities/split_triangle.h"
 #include "utilities/split_tetrahedra.h"
 #include "geometries/triangle_2d_3.h"
 #include "geometries/tetrahedra_3d_4.h"
