+++ kratos/utilities/cutting_utility.h
@@ -46,7 +46,7 @@
 #include "includes/mesh.h"
 #include "utilities/math_utils.h"
 #include "utilities/split_tetrahedra.h"
-#include "utilities/split_triangle.c"
+#include "utilities/split_triangle.h"
 #include "geometries/tetrahedra_3d_4.h"
 #include "geometries/triangle_3d_3.h"
 #include "spatial_containers/spatial_containers.h"
