+++ kratos/utilities/split_triangle.h
